Identify disk for volume by 'extra' field
Nailgun stores disk and volumes data separately: disk data in node's 'meta'
json blob, and volumes data in attributes for the node. User can't access the
attributes directly via API. Only node metadata is accessible.

In combination with known udev issue that breaks identification of devices by
PCI path [1], it can create a situation when disk identified differently in
volumes and in disk dicts.

This happens when during reboot from bootstrap to operating system after
provisioning udev changes the by-path linking of disks. The system ends up in a
situation like this:

  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0 -> ../../sdb
  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0-part1 -> ../../sdb1
  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0-part2 -> ../../sdb2
  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0-part3 -> ../../sda3
  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0-part4 -> ../../sda4
  /dev/disk/by-path/pci-0000:00:01.1-scsi-0:0:0:0-part5 -> ../../sda5

This, in turn, causes KeyError exception when user tries to restore the disk
data on the node for reinstallation, or after moving to another environment.

Change-Id: I26bc2fc3cc01c3356710aac98a69c3c1a653c741
Closes-bug: 1477604